requirements.txt
anyio==2.2.0
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
configparser==5.0.2
coverage==5.5
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.3
flake8==3.9.0
flake8-annotations==2.6.1
flake8-docstrings==1.6.0
flake8-import-order==0.18.1
gitdb==4.0.6
GitPython==3.1.14
idna==2.10
imageio==2.9.0
imgaug==0.4.0
iniconfig==1.1.1
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.5.1
jupyterlab==3.0.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
jupyterlab-vim==0.13.4
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.2
mccabe==0.6.1
mistune==0.8.4
mypy==0.812
mypy-extensions==0.4.3
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
notebook==6.3.0
numpy==1.20.1
opencv-python==4.5.1.48
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
pluggy==0.13.1
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.15.6
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pydocstyle==6.0.0
pyflakes==2.3.1
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.2
pytest-cov==2.11.1
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.2
Send2Trash==1.5.0
sentry-sdk==1.0.0
Shapely==1.7.1
shortuuid==1.0.1
six==1.15.0
sklearn==0.0
smmap==3.0.5
sniffio==1.2.0
snowballstemmer==2.1.0
subprocess32==3.5.4
terminado==0.9.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.3.17
toml==0.10.2
torch==1.8.1
torchvision==0.9.1
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
wandb==0.10.23
wcwidth==0.2.5
webencodings==0.5.1
